+++ aquamark.php
@@ -1,18 +1,18 @@
 <?php
 
 /**
- * Plugin Name:       Free Watermarks
- * Plugin URI:        https://github.com/williarin/free-watermarks
- * Description:       The best free watermarking plugin ever created. Add watermarks to your images in the WordPress media library.
+ * Plugin Name:       AquaMark
+ * Plugin URI:        https://github.com/williarin/aquamark
+ * Description:       Add a custom watermark to your images in the WordPress media library with powerful controls and blending modes.
  * Version:           1.0.0
  * Author:            William Arin
  * Author URI:        https://github.com/williarin
  * License:           GPL-2.0-or-later
  * License URI:       https://www.gnu.org/licenses/gpl-2.0.html
- * Text Domain:       free-watermarks
+ * Text Domain:       aquamark
  */
 
-use Williarin\FreeWatermarks\Plugin;
+use Williarin\AquaMark\Plugin;
 
 if (!defined('WPINC')) {
     die;